rename remote methods
